cleanup, fixed in release for LRC, added support for FB (wip)

+++ lib/uc_release.sh
@@ -83,7 +83,8 @@ ci_job_release() {
         execute mv ${WORKSPACE}/label.txt ${workspace}/bld/bld-lfs-release/oldLabel.txt
     else
         # TODO: demx2fk3 2014-08-08 this should be an error message.
-        touch ${workspace}/bld/bld-lfs-release/oldLabel.txt
+        local basedOn=$(getConfig LFS_PROD_uc_release_based_on)
+        echo ${basedOn} > ${workspace}/bld/bld-lfs-release/oldLabel.txt
     fi
 
     export LFS_PROD_RELEASE_CURRENT_TAG_NAME=$(cat ${workspace}/bld/bld-lfs-release/label.txt)
@@ -189,6 +190,8 @@ extractArtifactsOnReleaseShare() {
                 destination=${resultBuildShareLinuxKernel}/${labelName}
             fi
             executeOnMaster chmod u+w ${resultBuildShareLinuxKernel}/
+            warning "skipping kernel sources - copy to share is broken at the moment"
+            continue
         fi
 
         info "copy ${basename} to buildresults share ${destination}"
@@ -201,7 +204,7 @@ extractArtifactsOnReleaseShare() {
             execute rsync -av --exclude=.svn ${workspace}/bld/${basename}/. ${server}:${destination}
             touch ${destination}
         else
-            info "storing artifacts on share is disabled in config"
+            warning "storing artifacts on share is disabled in config"
         fi
     done
 
@@ -241,7 +244,6 @@ sendReleaseNote() {
     mustHaveValue "${osTagName}" "next os tag name"
     mustHaveValue "${oldReleaseTagName}" "old release tag name"
 
-
     # create the os or uboot release note
     info "new release label is ${releaseTagName} based on ${oldReleaseTagName}"
     _createLfsOsReleaseNote ${buildJobName} ${buildBuildNumber}
@@ -264,9 +266,6 @@ sendReleaseNote() {
         _createLfsRelReleaseNoteXml  ${releaseTagName} ${workspace}/rel/releasenote.xml
         _createReleaseInWorkflowTool ${releaseTagName} ${workspace}/rel/releasenote.xml
         _uploadToWorkflowTool        ${releaseTagName} ${workspace}/rel/releasenote.xml
-        _uploadToWorkflowTool        ${releaseTagName} ${workspace}/os/os_releasenote.xml
-        _uploadToWorkflowTool        ${releaseTagName} ${workspace}/os/releasenote.txt
-        _uploadToWorkflowTool        ${releaseTagName} ${workspace}/os/changelog.xml
 
         execute cp -f ${workspace}/rel/releasenote.xml ${workspace}/bld/bld-lfs-release/lfs_rel_releasenote.xml
     fi
@@ -274,10 +273,10 @@ sendReleaseNote() {
     if [[ ${canSendReleaseNote} ]] ; then
         info "send release note"
         execute ${LFS_CI_ROOT}/bin/sendReleaseNote  -r ${workspace}/os/releasenote.txt \
-                                                    -t ${releaseTagName}                 \
+                                                    -t ${releaseTagName}               \
                                                     -f ${LFS_CI_ROOT}/etc/file.cfg
     else
-        info "sending the release note is disabled in config"
+        warning "sending the release note is disabled in config"
     fi
 
     for file in ${workspace}/bld/bld-lfs-release/* ; do
@@ -303,6 +302,12 @@ _createReleaseInWorkflowTool() {
     local wftApiKey=$(getConfig WORKFLOWTOOL_api_key)
     local wftCreateRelease=$(getConfig WORKFLOWTOOL_url_create_release)
     local curl="curl -k ${wftCreateRelease} -F access_key=${wftApiKey}" 
+
+    local canCreateReleaseInWorkflowTool=$(getConfig LFS_CI_uc_release_can_create_release_in_wft)
+    if [[ ! ${canCreateReleaseInWorkflowTool} ]] ; then
+        warning "creating release in WFT is disabled via config"
+        return
+    fi
 
     # check, if wft already knows about the release
     local update=""
@@ -335,6 +340,12 @@ _uploadToWorkflowTool() {
     local wftApiKey=$(getConfig WORKFLOWTOOL_api_key)
     local wftUploadAttachment=$(getConfig WORKFLOWTOOL_url_upload_attachment)
     local curl="curl -k ${wftUploadAttachment}/${tagName} -F access_key=${wftApiKey} "
+
+    local canCreateReleaseInWorkflowTool=$(getConfig LFS_CI_uc_release_can_create_release_in_wft)
+    if [[ ! ${canCreateReleaseInWorkflowTool} ]] ; then
+        warning "uploading release note / file in WFT is disabled via config"
+        return
+    fi
 
     if [[ ! -s ${fileName} ]] ; then
         debug "file ${fileName} is empty"
@@ -383,6 +394,11 @@ _validateReleaseNoteXml() {
     local wftApiKey=$(getConfig WORKFLOWTOOL_api_key)
     local wftReleaseNoteValidate=$(getConfig WORKFLOWTOOL_url_releasenote_validation)
     local wftReleaseNoteXsd=$(getConfig WORKFLOWTOOL_url_releasenote_xsd)
+
+#    if [[ ! ${canCreateReleaseInWorkflowTool} ]] ; then
+#        warning "validating release note xml in WFT is disabled via config"
+#        return
+#    fi
 
     if [[ ${wftReleaseNoteXsd} ]] ; then
         # local check first
@@ -449,13 +465,14 @@ _createLfsOsReleaseNote() {
 
     export tagName=${LFS_PROD_RELEASE_CURRENT_TAG_NAME}
 
-    ${LFS_CI_ROOT}/bin/getReleaseNoteContent -t ${LFS_PROD_RELEASE_CURRENT_TAG_NAME} \
-                                             -f ${LFS_CI_ROOT}/etc/file.cfg > releasenote.txt
+    debug "${LFS_CI_ROOT}/bin/getReleaseNoteContent -t ${LFS_PROD_RELEASE_CURRENT_TAG_NAME} > releasenote.txt"
+    ${LFS_CI_ROOT}/bin/getReleaseNoteContent -t ${LFS_PROD_RELEASE_CURRENT_TAG_NAME} > releasenote.txt
     mustBeSuccessfull "$?" "getReleaseNoteContent"
     rawDebug ${workspace}/os/releasenote.txt
 
     export type=OS
 
+    debug " ${LFS_CI_ROOT}/bin/getReleaseNoteXML -t ${LFS_PROD_RELEASE_CURRENT_TAG_NAME} -o ${LFS_PROD_RELEASE_PREVIOUS_TAG_NAME} -r http://foobar/asdf/ -f ${LFS_CI_ROOT}/etc/file.cfg > releasenote.xml"
     ${LFS_CI_ROOT}/bin/getReleaseNoteXML -t ${LFS_PROD_RELEASE_CURRENT_TAG_NAME}  \
                                          -o ${LFS_PROD_RELEASE_PREVIOUS_TAG_NAME} \
                                          -r http://foobar/asdf/ \
@@ -589,6 +606,7 @@ createTagOnSourceRepository() {
     fi
 
     mustExistBranchInSubversion ${svnUrlOs}/branches ${branch}
+    mustExistBranchInSubversion ${svnUrlOs}/tags ${branch}
 
     for revisionFile in ${workspace}/rev/* ; do
         [[ -e ${revisionFile} ]] || continue
@@ -646,7 +664,7 @@ createTagOnSourceRepository() {
         info "branch ${branch} no longer required, removing branch"
         svnRemove -m removing_branch_for_production ${svnUrlOs}/branches/${branch} 
     else
-        info "creating a source tag is disabled in config"
+        warning "creating a source tag is disabled in config"
     fi
 
     info "tagging done"
@@ -703,6 +721,8 @@ createReleaseTag() {
 
     # check for the branch
     info "svn repos url is ${svnUrl}/branches/${branch}"
+    mustExistBranchInSubversion ${svnUrl} tags 
+    mustExistBranchInSubversion ${svnUrl} branches 
     shouldNotExistsInSubversion ${svnUrl}/tags/ "${osReleaseLabelName}"
 
     if ! existsInSubversion ${svnUrl}/branches ${branch} ; then
@@ -741,7 +761,7 @@ createReleaseTag() {
         echo "BTSPS-1657 IN rh: DESRIPTION: NOJCHK : create new tag ${osReleaseLabelName}" > ${logMessage}
         svnCopy -F ${logMessage} ${svnUrl}/branches/${branch} ${svnUrl}/tags/${osReleaseLabelName}
     else
-        info "creating the release tag is disabled in config"
+        warning "creating the release tag is disabled in config"
     fi
 
     info "tag created."
@@ -796,6 +816,7 @@ createProxyReleaseTag() {
     info "using ${deliverySvnUrl}"
 
     mustExistBranchInSubversion ${proxySvnUrl} branches 
+    mustExistBranchInSubversion ${proxySvnUrl} tags 
     mustExistBranchInSubversion ${proxySvnUrl}/branches ${branchName}
 
     svn pg svn:externals ${deliverySvnUrl}/tags/${relTagName} > ${externals}
@@ -815,7 +836,7 @@ createProxyReleaseTag() {
         echo "BTSPS-1657 IN rh: DESRIPTION: NOJCHK : creating new proxy tag ${tagName}" > ${logMessage}
         svnCopy -F ${logMessage} ${proxySvnUrl}/branches/${branchName} ${proxySvnUrl}/tags/${relTagName}
     else
-        info "creating a proxy tag is disabled in config"
+        warning "creating a proxy tag is disabled in config"
     fi
 
     info "creating proxy tag done."
@@ -834,7 +855,7 @@ updateDependencyFiles() {
 
     local workspace=$(getWorkspaceName)
     mustHaveWorkspaceName
-    mustHaveWorkspaceWithArtefactsFromUpstreamProjects ${jobName} ${buildNumber}
+    mustHaveWorkspaceWithArtefactsFromUpstreamProjects ${jobName} ${buildNumber} "externalComponents"
 
     local componentsFile=$(createTempFile)
     sort -u ${workspace}/bld/bld-externalComponents-*/usedRevisions.txt > ${componentsFile}
@@ -843,6 +864,7 @@ updateDependencyFiles() {
     local oldReleaseLabelName=${LFS_PROD_RELEASE_PREVIOUS_TAG_NAME}
     local canCommitDependencies=$(getConfig LFS_CI_uc_release_can_commit_depencencies)
 
+    info "using values: old value: ${oldReleaseLabelName} new value: ${releaseLabelName}"
     while read name url rev ; do
 
         case ${name} in
@@ -854,23 +876,25 @@ updateDependencyFiles() {
         debug "from file: ${name} ${url} ${rev}"
 
         execute rm -rf ${workspace}/${name}
-        svnCheckout --ignore-externals ${url} ${workspace}/${name}
+        svnCheckout --depth=immediates --ignore-externals ${url} ${workspace}/${name}
         local dependenciesFile=${workspace}/${name}/Dependencies
-        [[ ! -e ${dependenciesFile} ]] ; continue
-        
-        debug "update ${name}/Dependencies"
-        # TODO: demx2fk3 2014-07-14 we should also check the word boundary
-        # perl -p -i -e 's/\b${oldReleaseLabelName}\b/${releaseLabelName}/g' ${name}/Dependencies
-        perl -p -i -e 's/${oldReleaseLabelName}/${releaseLabelName}/g' ${dependenciesFile}
-        mustBeSuccessfull "$?" "perl -pie s/... ${dependenciesFile}"
 
+        if [[ ! -e ${dependenciesFile} ]] ; then
+            info "file ${dependenciesFile} does not exist"
+            continue
+        fi
+        
+        info "update ${name}/Dependencies";
+        execute perl -p -i -e "s/\b${oldReleaseLabelName}\b/${releaseLabelName}/g" ${dependenciesFile}
         svnDiff ${dependenciesFile}
+
         if [[ ${canCommitDependencies} ]] ; then 
+            info "running svn commit"
             local logMessage=$(createTempFile)
             echo "BTSPS-1657 IN psulm: DESCRIPTION: set Dependencies for Release ${releaseLabelName} r${rev} NOJCHK"  > ${logMessage}
             svnCommit -F ${logMessage} ${dependenciesFile}
         else
-            info "committing of dependencies is disabled in config"
+            warning "committing of dependencies is disabled in config"
         fi
